refactor HTTPWorker into consistencyworker
